+++ podaac/subsetter/run_subsetter.py
@@ -83,19 +83,38 @@ def parse_args(args: list) -> tuple:
         action='store_true',
         help='To pixel cut based the lon lat on the bounding box'
     )
+    parser.add_argument(
+        '--vertical-var',
+        type=str,
+        default=None,
+        help='Name of the vertical variable to subset by (e.g., depth, altitude)'
+    )
+    parser.add_argument(
+        '--vertical-min',
+        type=float,
+        default=None,
+        help='Minimum value for vertical subsetting (inclusive)'
+    )
+    parser.add_argument(
+        '--vertical-max',
+        type=float,
+        default=None,
+        help='Maximum value for vertical subsetting (inclusive)'
+    )
 
     args = parser.parse_args(args=args)
     bbox = np.array([[args.bbox[0], args.bbox[2]], [args.bbox[1], args.bbox[3]]])
 
     return args.input_file, args.output_file, bbox, args.variables, \
-        args.min_time, args.max_time, args.cut, args.shapefile, args.pixel_subset
+        args.min_time, args.max_time, args.cut, args.shapefile, args.pixel_subset, \
+        args.vertical_var, args.vertical_min, args.vertical_max
 
 
 def run_subsetter(args: list) -> None:
     """
     Parse arguments and run subsetter on the specified input file
     """
-    input_file, output_file, bbox, variables, min_time, max_time, cut, shapefile, pixel_subset = parse_args(args)
+    input_file, output_file, bbox, variables, min_time, max_time, cut, shapefile, pixel_subset, vertical_var, vertical_min, vertical_max = parse_args(args)
 
     logging.info('Executing subset on %s...', input_file)
     subset.subset(
@@ -108,7 +127,10 @@ def run_subsetter(args: list) -> None:
         max_time=max_time,
         origin_source=os.path.basename(input_file),
         shapefile=shapefile,
-        pixel_subset=pixel_subset
+        pixel_subset=pixel_subset,
+        vertical_var=vertical_var,
+        vertical_min=vertical_min,
+        vertical_max=vertical_max
     )
     logging.info('Subset complete. Result in %s', output_file)